Implement willDisplayCardsPublisher and send a pixel when addToDock is presented
